Upgrade to pass tests on v1.1.0-dev
I made minimal changes to the code; enough to pass the existing test
suite.

A couple of details:

* Define the function `eye` which was removed in Julia v1.0. This is a
stop-gap. It is probably best to rewrite the code that calls `eye`.

* Prefer `import` and fully-qualified symbols to `using` and `import`ing
symbols. I only did this in code that required upgrading. In particular,
in v1.0 `LinearAlgebra` is no longer imported into `Base`. I use fully-
qualified names for symbols in `LinearAlgebra`.