Support AbstractArray types
Replace `Matrix` and `Vector` by `AbstractMatrix` and `AbstractVector` in
function signatures.

The modified signatures are tested by repeating one of the testsets
with `eye` replaced by a functionally-equivalent `alteye`. I have also
added `@time` macros before the test sets. This is a good way to spot
gross efficiency regressions. The timing for `alteye` is faster than
that for `eye`. But, the trial with `eye` comes first, so the majority
of reported time represents jit compilation. In fact, `alteye` is slower,
at least for these tests. But, the main point is to test `AbstractArray`
types.